\name{chain-class}
\Rdversion{1.1}
\docType{class}
\alias{chain-class}

\title{Class \code{"chain"}}
\description{
The chain class defines how the chain data is stored for use in the Error_Function 
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("chain", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{size}:}{Object of class \code{"integer"} ~~ }
    \item{\code{prices}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{types}:}{Object of class \code{"character"} ~~ }
    \item{\code{strikes}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{taus}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{s}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{weights}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "chain" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("chain")
}
\keyword{classes}
