\name{gpusvcalibration-package}
\alias{gpusvcalibration-package}
\alias{gpusvcalibration}
\docType{package}
\title{
R package to support fast calibration of stochastic volatility models for option pricing using GPUs 
}
\description{
The purpose of this package is to demonstrate how stochastic volatility models can be quickly calibrated with standard constrained optimization routines such as those available in NLoptr. This package provides an interface called ErrorFunction which efficiently offloads the pricing computations for the chain on the gpu.
}
\details{
\tabular{ll}{
Package: \tab gpusvcalibration\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-26\cr
License: \tab BSD (clause 3)\cr
Depends: \tab methods\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
